Predefine the value of CMAKE_INSTALL_INCLUDEDIR
Defining the value of this cache variable before including the
GNUInstallDirs module will avoid the issue of this library causing the
implicit inclusion of other libraries from the system include path if
this library is consumed from a system installation.

If the user already provided a value for CMAKE_INSTALL_INCLUDEDIR, then
this will be a noop due to it being a cache variable.

The value of "include/<project-name>" is always safe, because the CMake
defined of value of CMAKE_INSTALL_INCLUDEDIR has never been anything
other than "include".