Fix variables in the superbuild CML
CMAKE_INSTALL_INCLUDEDIR is provided only because in the CI workflow
the project is also installed and the root CML sets the variable only
as the top level project.

But that is okay, because superbuild CMLs are exactly for serving as
containment for gruesome details like that, so the root CML can stay an
almost* pure description of the project.

*: some things aren't purely descriptions of the project, but that is
merely to provide some convenience features to people who vendor the
code. This is especially important since people are using FetchContent
more and more.