Introduce the developer mode variable
This variable will be used to cleanly separate parts of the project CML
that are intended for the developers of the project from the parts that
are also necessary for the consumers of the project, such as other
developers using the project as a dependency or package maintainers
building the project.

This variable will allow shedding the superbuild CML and the use of a
single CML as an entrypoint for the build.