Add .deb packaging example
Add files and scripts necessary to build .deb packages that pass the
lintian checks.

There is a hack in config.cmake that relies on an implementation detail
to suppress generating postinst and postrm debian package files that
raise warnings in lintian. These files can't be disabled by any
documented option.